Fix dtype of input image

+++ setup_predictor.py
@@ -6,7 +6,7 @@ import os
 
 
 def setup_predictor(model_path, vgg19_path):
-    images_pl = tf.placeholder(tf.float32, name='images', shape=(1, 64, 64, 3))
+    images_pl = tf.placeholder(tf.uint8, name='images', shape=(1, 64, 64, 3))
     configs_pl = tf.placeholder(tf.float32, name='configs', shape=(1, 10))
 
     with tf.variable_scope('model', reuse=None):
@@ -25,7 +25,7 @@ def setup_predictor(model_path, vgg19_path):
     saver.restore(sess, model_path)
 
     def predictor_func(images=None, robot_configs=None):
-        feed_image = images.astype(np.float32).reshape((1, 64, 64, 3))
+        feed_image = images.astype(np.uint8).reshape((1, 64, 64, 3))
         feed_config = robot_configs.astype(np.float32).reshape((1, 10))
 
         feed_dict = {
